// packages/@fluentui-vue/next/src/components/SwatchColorPicker/ColorPickerGridCell.types.ts
export interface IColorCellProps {
  /**
   * Arbitrary unique string associated with this option
   */
  id: string

  /**
   * Tooltip and aria label for this item
   */
  label?: string

  /**
   * The CSS-compatible string to describe the color
   */
  color: string

  /**
   * Index for this option
   */
  index?: number
}

/**
 * {@docCategory SwatchColorPicker}
 */
export interface IColorPickerGridCellStyleProps {
  /**
   * Theme to apply to the cell.
   */
  theme: ITheme

  /**
   * Whether the component is disabled or not.
   */
  disabled?: boolean

  /**
   * Whether the cell is currently selected or not.
   */
  selected?: boolean

  /**
   * Whether the svg color element should be rendered as a circle or not.
   */
  circle?: boolean

  /**
   * Whether the color being rendered is white or not. If it is white we show a border around it.
   */
  isWhite?: boolean

  /**
   * The height of this cell, in pixels.
   */
  height?: number

  /**
   * The width of this cell, in pixels.
   */
  width?: number

  /**
   * The width of the border indicating a hovered or selected cell, in pixels.
   */
  borderWidth?: number
}

/**
 * {@docCategory SwatchColorPicker}
 */
export interface IColorPickerGridCellStyles {
  /**
   * Style to apply to a colorCell in the color picker.
   */
  colorCell: IStyle

  /**
   * Style to apply to the svg element that renders the color.
   */
  svg: IStyle
}